DISMTools 0.6.1 Preview 3 (#216)
* Update assembly information

* [PE Helper] Update HotInstall (01/25/2025)

Updated HotInstall to its latest commit: https://github.com/CodingWonders/DT-HotInstall/commit/6e87e16d7c01a5ed3b65679341d158afcbe03e00

* [Updater] Add error information

* [PE Helper] Update HotInstall (01/25/2025, part 2)

Updated HotInstall to its latest commit: https://github.com/CodingWonders/DT-HotInstall/commit/af626ff988545d5783da3e8aa6b29070ab088772

* [Unattended answer file] Block "None" user account name

* [PE Helper] Detect if image index is valid

* [PE Helper] Add support for Ventoy drives

* [ISO Creator] Remove redundant label

* [PE Helper] Update HotInstall (01/29/2025)

Update HotInstall to its latest commit: https://github.com/CodingWonders/DT-HotInstall/commit/43ce929fb71a9b215736d49c1caaa518d8515a69

NOTE: this adds language support to HotInstall for the following languages, based on ISO 639-2 language code representations: English, Spanish (Spain). Unsupported languages will default to English for the HotInstall installer

* [PE Helper] Update HotInstall (01/30/2025)

Update HotInstall to its latest commit: https://github.com/CodingWonders/DT-HotInstall/commit/f5df6f0727cc5e54005f42a859f93c88a80ea1f7

* [PE Helper] Update HotInstall (01/31/2025)

Update HotInstall to its latest commit: https://github.com/CodingWonders/DT-HotInstall/commit/e6031580fb2ec60f8c47c49272637c082cefc870

* [Unattended answer file] UnattendGen Nightly

* [Fix] Switch to manual size for label

* [Build] Remove dependency for speedier CI

* [Build] Remove dependency for speedier CI (part 2)

* [Enhance] Make progress panel log switchers more appealing

* [PE Helper] Show visual progress of DIM driver addition

* [PE Helper] Hide progress once completed

* [PE Helper] Hide progress once completed (part 2)

* [PE Helper] Change progress output of driver installation

* [PE Helper] Update HotInstall (02/02/2025)

Update HotInstall to its latest commit: https://github.com/CodingWonders/DT-HotInstall/commit/d843f50d732dd421d5337eb3bc2fa69c427fc0f7

* [PE Helper] Update HotInstall (02/02/2025, part 2)

Update HotInstall to its latest commit: https://github.com/CodingWonders/DT-HotInstall/commit/bc05ccd7eb4ff15cbeefb3e09df122aca49abaaa

* [DynaLog] Change file length type to long

* [ISO Creator] Detect ADKs and WIMs for architectures

* [PE Helper] Fix compat issues with ADK 10.0.26100.2454+

* [PE Helper] Add option for signed boot binaries

* [Library] Update Scintilla.NET to 5.6.6

* [Unattended answer file & PE Helper] Update vers.

- Updated UnattendGen (v0.6.1.2521)
- Updated HotInstall to remove test build flag

* [REL] Update What's New section, new Update Info files

+++ Utilities/DynaLog.vb
@@ -86,7 +86,7 @@ Public Class DynaLog
                 Debug.WriteLine("Creating log directory...")
                 Directory.CreateDirectory(Application.StartupPath & "\logs")
             End If
-            Dim FileLength As Integer = 0
+            Dim FileLength As Long = 0
             If File.Exists(Application.StartupPath & "\logs\DT_DynaLog.log") Then
                 FileLength = New FileInfo(Application.StartupPath & "\logs\DT_DynaLog.log").Length
             End If
